File was in wrong dir
